Add files via upload
The top five destination list file is a Java file which was created to showcase the top five destinations the class individually chose to represent our first iteration of the "SNHU Travel Agency" recommendations for travelers booking through our example service. The SlideShow.jar file is a Java file which is an updated iteration of the previous top five destination list, which was an additional assignment that covered an example of if a theoretical Product Owner requested a change to the code in response to updated business needs.